Refactor according payment succeed hook
