+++ packages/realm-server/billing/billing-queries.ts
@@ -104,6 +104,19 @@ export async function getPlanByStripeId(
   } as Plan;
 }
 
+export async function updateUserStripeCustomerId(
+  dbAdapter: DBAdapter,
+  userId: string,
+  stripeCustomerId: string,
+) {
+  await query(dbAdapter, [
+    `UPDATE users SET stripe_customer_id = `,
+    param(stripeCustomerId),
+    ` WHERE matrix_user_id = `,
+    param(userId),
+  ]);
+}
+
 export async function getUserByStripeId(
   dbAdapter: DBAdapter,
   stripeCustomerId: string,
