+++ packages/realm-server/billing/stripe-webhook-handlers/index.ts
@@ -1,6 +1,6 @@
 import { DBAdapter } from '@cardstack/runtime-common';
 import { handlePaymentSucceeded } from './payment-succeeded';
-import { handleCheckoutSessionCompleted } from './subscribe';
+import { handleCheckoutSessionCompleted } from './checkout-session-completed';
 
 import Stripe from 'stripe';
 import { handleSubscriptionDeleted } from './subscription-deleted';
